sha3: rename impl_digest_trait to hasher
